+++ libs/client/triggers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shipfox/client-triggers
 
+## 31.0.1
+
+### Patch Changes
+
+- @shipfox/client-ui@31.0.1
+- @shipfox/client-projects@31.0.1
+- @shipfox/client-shell@31.0.1
+
 ## 31.0.0
 
 ### Patch Changes